refactor: buildPacket3()
